2.3.2618

+++ app/services/AssetsService.php
@@ -325,6 +325,11 @@ class AssetsService extends BaseApplicationComponent
 	 */
 	public function getFolderTreeBySourceIds($allowedSourceIds)
 	{
+		if (empty($allowedSourceIds))
+		{
+			return array();
+		}
+
 		$folders = $this->findFolders(array('sourceId' => $allowedSourceIds, 'order' => 'path'));
 		$tree = $this->_getFolderTreeByFolders($folders);
 
@@ -795,11 +800,12 @@ class AssetsService extends BaseApplicationComponent
 	/**
 	 * Delete a list of files by an array of ids (or a single id).
 	 *
-	 * @param $fileIds
+	 * @param array $fileIds
+	 * @param bool $deleteFile Should the file be deleted along the record. Defaults to true.
 	 *
 	 * @return AssetOperationResponseModel
 	 */
-	public function deleteFiles($fileIds)
+	public function deleteFiles($fileIds, $deleteFile = true)
 	{
 		if (!is_array($fileIds))
 		{
@@ -820,7 +826,11 @@ class AssetsService extends BaseApplicationComponent
 					'asset' => $file
 				)));
 
-				$source->deleteFile($file);
+				if ($deleteFile)
+				{
+					$source->deleteFile($file);
+				}
+
 				craft()->elements->deleteElementById($fileId);
 
 				// Fire an 'onDeleteAsset' event
@@ -1062,7 +1072,10 @@ class AssetsService extends BaseApplicationComponent
 				throw new Exception(Craft::t('That folder does not seem to exist anymore. Re-index the Assets source and try again.'));
 			}
 
-			if (!craft()->userSession->checkPermission($permission.':'.$folderModel->sourceId))
+			if (
+				!craft()->userSession->checkPermission($permission.':'.$folderModel->sourceId)
+				&&
+				!craft()->userSession->checkAuthorization($permission.':'.$folderModel->id))
 			{
 				throw new Exception(Craft::t('You don’t have the required permissions for this operation.'));
 			}
@@ -1324,6 +1337,7 @@ class AssetsService extends BaseApplicationComponent
 		if ($fileId)
 		{
 			$response->setDataItem('fileId', $fileId);
+			$response->setDataItem('filename', $theNewFile->filename);
 		}
 
 		return $response;
